Initial attempt to move routes out but fails
